+++ src/config.py
@@ -1,6 +1,13 @@
 import os
+import uuid
 
 BOT_TOKEN = os.environ.get("BOT_TOKEN", "")
+
+# HOOK_SERVER - ip or url of server where bot is serving
+HOOK_SERVER = os.environ.get("HOOK_SERVER", "")
+HOOK_PORT = os.environ.get("HOOK_PORT", 8080)
+# HOOK_URL - url of hook
+HOOK_URL = os.environ.get("HOOK_URL", str(uuid.uuid4()))
 
 # Week number from start of the year
 START_WEEK = 6
